+++ bundles/aero.minova.rcp.rcp/src/aero/minova/rcp/rcp/util/CreateButtonInDetailUtil.java
@@ -10,6 +10,7 @@ import org.eclipse.core.commands.ParameterizedCommand;
 import org.eclipse.e4.core.commands.ECommandService;
 import org.eclipse.e4.core.commands.EHandlerService;
 import org.eclipse.e4.core.di.annotations.Optional;
+import org.eclipse.e4.core.services.events.IEventBroker;
 import org.eclipse.e4.core.services.translation.TranslationService;
 import org.eclipse.e4.ui.model.application.MApplication;
 import org.eclipse.e4.ui.model.application.commands.MCommand;
@@ -77,6 +78,9 @@ public class CreateButtonInDetailUtil {
 
 	@Inject
 	MPart mPart;
+
+	@Inject
+	IEventBroker broker;
 
 	@Inject
 	@Optional
@@ -248,6 +252,10 @@ public class CreateButtonInDetailUtil {
 						} else {
 							// Auch in Methode createParameters() anpassen!!
 							System.err.println("Event vom Typ " + o.getClass() + " für Buttons noch nicht implementiert!");
+						}
+
+						if (onclick.isReloadDetail()) {
+							broker.post(Constants.BROKER_RELOADFIELDS, null);
 						}
 					}
 				}