+++ bundles/aero.minova.rcp.rcp/src/aero/minova/rcp/rcp/widgets/SectionGrid.java
@@ -231,7 +231,7 @@ public class SectionGrid {
 		loadState();
 
 		for (int i = 0; i < dataTable.getColumnCount(); i++) {
-			originalReadOnlyColumns.put(i, dataTable.getColumns().get(i).isReadOnly());
+			originalReadOnlyColumns.put(i, dataTable.getColumns().get(i).isReadOnly() || grid.isReadOnly());
 			originalRequiredColumns.put(i, dataTable.getColumns().get(i).isRequired());
 		}
 	}
@@ -384,6 +384,8 @@ public class SectionGrid {
 		selectionLayer = new SelectionLayer(columnHideShowLayer);
 
 		// Delete Button updaten (nur aktiviert, wenn eine Zelle gewählt ist)
+		deleteToolItemAccessor.setCanBeEnabled(false); // Erst mal Löschen deaktivieren
+		deleteToolItemAccessor.updateEnabled();
 		selectionLayer.addLayerListener(event -> {
 			if (deleteToolItemAccessor != null && event instanceof ISelectionEvent) {
 				deleteToolItemAccessor.setCanBeEnabled(selectionLayer.getSelectedCellPositions().length > 0);
@@ -446,7 +448,7 @@ public class SectionGrid {
 		getNatTable().setConfigRegistry(configRegistry);
 		getNatTable().addConfiguration(new DefaultNatTableStyleConfiguration());
 		getNatTable().addConfiguration(new SingleClickSortConfiguration());
-		gridConfiguration = new MinovaGridConfiguration(dataTable.getColumns(), mGrid.getFields(), dataService, translationService);
+		gridConfiguration = new MinovaGridConfiguration(dataTable.getColumns(), mGrid.getFields(), dataService, translationService, grid.isReadOnly());
 		getNatTable().addConfiguration(gridConfiguration);
 		columnHideShowLayer.hideColumnPositions(gridConfiguration.getHiddenColumns());